+++ code-types/TypeConverter.js
@@ -1,5 +1,5 @@
 "use strict";
-/** TypeConverter - functions for converting source code data types between different languages
+/** Helper functions for converting source code data types between different languages
  * @author TeamworkGuy2
  */
 var TypeConverter;